Move object classes to separate files
Not sure if this is even the right way to do this. I'm kinda winging this, creating a project structure the same way I would do if I'm working on a C# app.
