+++ scripts/objects/page.mjs
@@ -1,0 +1,86 @@
+// Page object.
+export class Page {
+  constructor(cookies) {
+    this.cookies = cookies;
+  }
+
+  // Private methods.
+  #setBetweenRange(minNumber, maxNumber) {
+    document.getElementById(
+      "between"
+    ).innerHTML = `(Between ${minNumber} and ${maxNumber})`;
+  }
+
+  #setScore(currentScore) {
+    document.getElementById("score").innerHTML = `${currentScore}`;
+  }
+
+  #setTotalScore(totalScore) {
+    if (totalScore < 1) {
+      document.getElementById("total-score").innerHTML = `0`;
+    } else {
+      document.getElementById("total-score").innerHTML = `${totalScore.toFixed(
+        1
+      )}`;
+    }
+  }
+
+  #setShortcutNumbers(minNumber, maxNumber) {
+    document.getElementById("min-number").innerHTML = `${minNumber}`;
+    document.getElementById("max-number").innerHTML = `${maxNumber}`;
+    document.getElementById("mid-number").innerHTML = `${Math.ceil(
+      maxNumber / 2
+    )}`;
+  }
+
+  // Public methods.
+  setValuesOnPage(player) {
+    this.#setBetweenRange(player.minNumber, player.maxNumber);
+    this.#setScore(player.currentScore);
+    this.#setTotalScore(player.totalScore);
+    this.#setShortcutNumbers(player.minNumber, player.maxNumber);
+  }
+
+  resetScreen(player) {
+    player.isGuessed = false;
+    document.querySelector("body").style.background = "#222";
+    document.querySelector(".number").style.width = "15rem";
+    document.querySelector(".btn.again").style.display = "none";
+    document.querySelector(".btn.check").style.display = "inline-block";
+    document.querySelector(".number").innerHTML = "?";
+    document.querySelector("#message").innerHTML = "Start guessing...";
+  }
+
+  guessedNumber(secretNumber) {
+    document.querySelector("body").style.background = "#60b347";
+    document.querySelector(".number").style.width = "30rem";
+    document.querySelector(".btn.check").style.display = "none";
+    document.querySelector(".btn.again").style.display = "inline-block";
+    document.querySelector(".number").innerHTML = `${secretNumber}`;
+    this.cookies.setCookies();
+  }
+
+  updateOnClick(player, playerGuess, guessMessageElement) {
+    if (playerGuess == player.secretNumber) {
+      this.guessedNumber(player.secretNumber);
+      guessMessageElement.innerHTML =
+        "✅ You guessed the number! Changing number...";
+      player.totalScore += player.currentScore;
+      player.setSecretNumber();
+      player.isGuessed = true;
+    } else if (playerGuess > player.secretNumber) {
+      guessMessageElement.innerHTML = "☝ Too high...";
+      player.currentScore--;
+    } else if (playerGuess < player.secretNumber) {
+      guessMessageElement.innerHTML = "👇 Too low...";
+      player.currentScore--;
+    }
+  }
+
+  outOfScore() {
+    document.querySelector("body").style.background = "#ab263a";
+    document.querySelector(".btn.check").style.display = "none";
+    document.querySelector(".btn.again").style.display = "inline-block";
+    document.querySelector(".number").innerHTML = "X";
+  }
+}
