+++ scripts/objects/player.mjs
@@ -1,0 +1,25 @@
+// Player stats.
+export class Player {
+  constructor() {
+    this.minNumber = 1;
+    this.maxNumber = 5;
+
+    this.currentScore = Math.ceil(this.maxNumber / 2);
+    this.totalScore = 0.000001;
+
+    // TODO: Make secret number private to it can't be taken from the dev console.
+    // Easy fix but rn I'm too lazy to update it ;D
+    this.secretNumber =
+      Math.floor(Math.random() * this.maxNumber) + this.minNumber;
+    this.isGuessed = false;
+  }
+
+  resetScore() {
+    this.currentScore = Math.ceil(this.maxNumber / 2);
+  }
+
+  setSecretNumber() {
+    this.secretNumber =
+      Math.floor(Math.random() * this.maxNumber) + this.minNumber;
+  }
+}
