+++ scripts/script.mjs
@@ -1,193 +1,13 @@
-// Imports.
+// Object imports.
+import { Player } from "./objects/player.mjs";
+import { Page } from "./objects/page.mjs";
+
+// Builder imports.
 import { SimpleAlgorithmBuilder } from "./builders/simple-algorithm-builder.mjs";
 import { FreelanceGuesserBuilder } from "./builders/freelance-guesser-builder.mjs";
+
+// Cookie import.
 import { Cookie } from "./cookies/cookies.mjs";
-
-// Player stats.
-class Player {
-  constructor() {
-    this.minNumber = 1;
-    this.maxNumber = 5;
-
-    this.currentScore = Math.ceil(this.maxNumber / 2);
-    this.totalScore = 0.000001;
-
-    // TODO: Make secret number private to it can't be taken from the dev console.
-    // Easy fix but rn I'm too lazy to update it ;D
-    this.secretNumber =
-      Math.floor(Math.random() * this.maxNumber) + this.minNumber;
-    this.isGuessed = false;
-  }
-
-  resetScore() {
-    this.currentScore = Math.ceil(this.maxNumber / 2);
-  }
-
-  setSecretNumber() {
-    this.secretNumber =
-      Math.floor(Math.random() * this.maxNumber) + this.minNumber;
-  }
-}
-
-// Page object.
-class Page {
-  constructor(cookies) {
-    this.cookies = cookies;
-  }
-
-  // Private methods.
-  #setBetweenRange(minNumber, maxNumber) {
-    document.getElementById(
-      "between"
-    ).innerHTML = `(Between ${minNumber} and ${maxNumber})`;
-  }
-
-  #setScore(currentScore) {
-    document.getElementById("score").innerHTML = `${currentScore}`;
-  }
-
-  #setTotalScore(totalScore) {
-    if (totalScore < 1) {
-      document.getElementById("total-score").innerHTML = `0`;
-    } else {
-      document.getElementById("total-score").innerHTML = `${totalScore.toFixed(
-        1
-      )}`;
-    }
-  }
-
-  #setShortcutNumbers(minNumber, maxNumber) {
-    document.getElementById("min-number").innerHTML = `${minNumber}`;
-    document.getElementById("max-number").innerHTML = `${maxNumber}`;
-    document.getElementById("mid-number").innerHTML = `${Math.ceil(
-      maxNumber / 2
-    )}`;
-  }
-
-  // Public methods.
-  setValuesOnPage(player) {
-    this.#setBetweenRange(player.minNumber, player.maxNumber);
-    this.#setScore(player.currentScore);
-    this.#setTotalScore(player.totalScore);
-    this.#setShortcutNumbers(player.minNumber, player.maxNumber);
-  }
-
-  resetScreen(player) {
-    player.isGuessed = false;
-    document.querySelector("body").style.background = "#222";
-    document.querySelector(".number").style.width = "15rem";
-    document.querySelector(".btn.again").style.display = "none";
-    document.querySelector(".btn.check").style.display = "inline-block";
-    document.querySelector(".number").innerHTML = "?";
-    document.querySelector("#message").innerHTML = "Start guessing...";
-  }
-
-  guessedNumber(secretNumber) {
-    document.querySelector("body").style.background = "#60b347";
-    document.querySelector(".number").style.width = "30rem";
-    document.querySelector(".btn.check").style.display = "none";
-    document.querySelector(".btn.again").style.display = "inline-block";
-    document.querySelector(".number").innerHTML = `${secretNumber}`;
-    this.cookies.setCookies();
-  }
-
-  updateOnClick(player, playerGuess, guessMessageElement) {
-    if (playerGuess == player.secretNumber) {
-      this.guessedNumber(player.secretNumber);
-      guessMessageElement.innerHTML =
-        "✅ You guessed the number! Changing number...";
-      player.totalScore += player.currentScore;
-      player.setSecretNumber();
-      player.isGuessed = true;
-    } else if (playerGuess > player.secretNumber) {
-      guessMessageElement.innerHTML = "☝ Too high...";
-      player.currentScore--;
-    } else if (playerGuess < player.secretNumber) {
-      guessMessageElement.innerHTML = "👇 Too low...";
-      player.currentScore--;
-    }
-  }
-
-  outOfScore() {
-    document.querySelector("body").style.background = "#ab263a";
-    document.querySelector(".btn.check").style.display = "none";
-    document.querySelector(".btn.again").style.display = "inline-block";
-    document.querySelector(".number").innerHTML = "X";
-  }
-}
-
-// Upgrade object.
-export class Upgrade {
-  #descriptionHtmlTemplate() {
-    let htmlTemplate = `<div class="item">
-      <h2 class="item-name">${this.name}</h3>
-      <span class="item-info">
-        Owned: <span class="count">${this.ownedAmount}</span> 
-        // Score Per Second: <span class="score-per-second">${this.scorePerSecond}</span> 
-        // Current SPS: <span class="current-sps">${this.totalScorePerSecond}</span>
-      </span>
-      <p class="item-description">
-        ${this.description}
-      </p>
-    </div>`;
-
-    return htmlTemplate;
-  }
-
-  #buttonsHtmlTemplate() {
-    let htmlTemplate = `<button class="btn buy-item">Buy <span class="amount">1</span> <span class="price-symbol">🥇</span><span
-    class="buy-price">${this.buyPrice}</span></button>
-    <button class="btn upgrade-it">Upgrade It <span class="price-symbol">🥇</span><span
-        class="upgrade-price">${this.upgradePrice}</span></button>
-    <button class="btn sell-item">Sell <span class="amount">1</span> <span class="price-symbol">🥇</span><span
-        class="sell-price">${this.sellPrice}</span></button>`;
-
-    return htmlTemplate;
-  }
-
-  #insertUpgradeDescriptionHtml() {
-    const parentHtmlElement = document.getElementById(this.descriptionHtmlId);
-    parentHtmlElement.insertAdjacentHTML(
-      "afterbegin",
-      this.#descriptionHtmlTemplate()
-    );
-  }
-
-  #insertUpgradeButtonsHtml() {
-    const parentHtmlElement = document.getElementById(this.buttonsHtmlId);
-    parentHtmlElement.insertAdjacentHTML(
-      "afterbegin",
-      this.#buttonsHtmlTemplate()
-    );
-  }
-
-  #insertUnrevealedUpgradeHtml() {
-    const parentHtmlElement = document.getElementById(this.descriptionHtmlId);
-    parentHtmlElement.insertAdjacentHTML(
-      "afterbegin",
-      `<h2 style="margin-bottom: 2rem">????????</h2><p style="color: rgb(158, 158, 158)">?????????????????</p>`
-    );
-  }
-
-  insertUpgradeHtml() {
-    if (this.isRevealed) {
-      this.#insertUpgradeDescriptionHtml();
-      this.#insertUpgradeButtonsHtml();
-    } else {
-      this.#insertUnrevealedUpgradeHtml();
-    }
-  }
-
-  buyUpgrade(player, page) {
-    if (player.totalScore >= this.buyPrice) {
-      player.totalScore -= this.buyPrice;
-      this.ownedAmount++;
-      page.setValuesOnPage(player);
-    } else {
-      console.log("not enough score to buy");
-    }
-  }
-}
 
 // Button events function.
 const buttonEvents = function (player, page) {
